separate UniquenessSet code
